refactor: extract shared notebook-doc crate from runtimed
Eliminates WASM copy drift — both runtimed (daemon) and runtimed-wasm
(frontend) now depend on one source of truth for the Automerge notebook
document schema and operations.

Changes:
- git mv runtimed/src/notebook_doc.rs → notebook-doc/src/lib.rs (history preserved)
- Delete runtimed-wasm/src/notebook_doc.rs (was a diverging copy)
- File I/O methods (load_or_create, save_to_file) gated behind 'persistence' feature
- StreamOutputState moved from stream_terminal.rs to the shared crate
  (2-field struct, no external deps) — runtimed re-exports it
- NotebookDoc gains doc()/doc_mut() accessors for extension by consumers
- runtimed depends on notebook-doc with features=["persistence"]
- runtimed-wasm depends on notebook-doc (no features)

Tests: 23 notebook-doc (19 core + 4 persistence), 233 runtimed, all passing.

Closes the 'extract shared notebook-doc crate' item in #556.
