+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "crates/kernel-launch",
     "crates/kernel-env",
     "crates/notebook",
+    "crates/notebook-doc",
     "crates/tauri-jupyter",
     "crates/xtask",
     "crates/runtimed",
