+++ crates/runtimed/src/lib.rs
@@ -19,7 +19,7 @@ pub mod connection;
 pub mod daemon;
 pub mod inline_env;
 pub mod kernel_manager;
-pub mod notebook_doc;
+pub use notebook_doc;
 pub mod notebook_metadata;
 pub mod notebook_sync_client;
 pub mod notebook_sync_server;
